+++ back/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'P3'
+rootProject.name = 'p3'
