--- back/src/main/java/Dto/AuthResponseDto.java
@@ -1,13 +1,0 @@
-package Dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class AuthResponseDto {
-    @JsonProperty("token")
-    private String token;
-    private String tokenType = "Bearer ";
-
-    public AuthResponseDto(String token) {
-        this.token = token;
-    }
-}
