--- back/src/main/java/Dto/LoginDto.java
@@ -1,22 +1,0 @@
-package Dto;
-
-public class LoginDto {
-    private String email;
-    private String password;
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
