--- back/src/main/java/Model/UserEntity.java
@@ -1,89 +1,0 @@
-package Model;
-
-import jakarta.persistence.*;
-
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
-
-@Entity
-@Table(name = "USERS")
-public class UserEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private long id;
-    private String email;
-    private String name;
-    private String password;
-    private Date createdAt;
-    private Date updatedAt;
-
-    @ManyToMany(fetch = FetchType.EAGER, cascade = {
-            CascadeType.DETACH,
-            CascadeType.MERGE,
-            CascadeType.PERSIST,
-            CascadeType.REFRESH
-    })
-    @JoinTable(
-            name = "USERS_ROLES",
-            joinColumns = @JoinColumn(name = "user_id", referencedColumnName = "id"),
-            inverseJoinColumns = @JoinColumn(name = "role_id", referencedColumnName = "id")
-    )
-    private List<RoleEntity> roles = new ArrayList<>();
-
-    public UserEntity() {
-        this.createdAt = new Date();
-    }
-
-    public Date getCreatedAt() {
-        return createdAt;
-    }
-
-    public Date getUpdatedAt() {
-        return updatedAt;
-    }
-
-    public void setUpdatedAt(Date updatedAt) {
-        this.updatedAt = updatedAt;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public UserEntity setEmail(String email) {
-        this.email = email;
-        return this;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public UserEntity setName(String name) {
-        this.name = name;
-        return this;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public UserEntity setPassword(String password) {
-        this.password = password;
-        return this;
-    }
-
-    public List<RoleEntity> getRoles() {
-        return roles;
-    }
-
-    public UserEntity setRoles(List<RoleEntity> roles) {
-        this.roles = roles;
-        return this;
-    }
-}
