--- back/src/main/java/Persistence/MessageRepository.java
@@ -1,9 +1,0 @@
-package Persistence;
-
-import Model.MessageEntity;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MessageRepository extends CrudRepository<MessageEntity, Integer> {
-}
