--- back/src/main/java/Persistence/RoleRepository.java
@@ -1,12 +1,0 @@
-package Persistence;
-
-import Model.RoleEntity;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface RoleRepository extends CrudRepository<RoleEntity, Integer> {
-    Optional<RoleEntity> findByName(String name);
-}
