--- back/src/main/java/Persistence/UserRepository.java
@@ -1,13 +1,0 @@
-package Persistence;
-
-import Model.UserEntity;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends CrudRepository<UserEntity, Integer> {
-    Optional<UserEntity> findByEmail(String email);
-    Boolean existsByEmail(String email);
-}
