+++ back/src/main/java/com/p3/controller/FileController.java
@@ -1,0 +1,37 @@
+package com.p3.controller;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.core.io.Resource;
+import org.springframework.core.io.UrlResource;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.io.IOException;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+
+@RequestMapping("api/files")
+@RestController
+public class FileController {
+    @Value("${assets.savePath}")
+    private String assetsPath;
+
+    @GetMapping("/file/{filename}")
+    public ResponseEntity<Resource> getFile(@PathVariable String filename) throws IOException {
+        Path filePath = Paths.get(assetsPath).resolve(filename);
+        Resource resource = new UrlResource(filePath.toUri());
+
+        if (!resource.exists()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        return ResponseEntity.ok()
+                .contentType(MediaType.IMAGE_JPEG)
+                .body(resource);
+    }
+}
