+++ back/src/main/java/com/p3/dto/AuthResponseDto.java
@@ -1,0 +1,12 @@
+package com.p3.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class AuthResponseDto {
+    @JsonProperty("token")
+    private String token;
+
+    public AuthResponseDto(String token) {
+        this.token = token;
+    }
+}
