+++ back/src/main/java/com/p3/dto/LoginDto.java
@@ -1,0 +1,22 @@
+package com.p3.dto;
+
+public class LoginDto {
+    private String email;
+    private String password;
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
