+++ back/src/main/java/com/p3/dto/RegisterDto.java
@@ -1,0 +1,31 @@
+package com.p3.dto;
+
+public class RegisterDto {
+    private String email;
+    private String username;
+    private String password;
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
