+++ back/src/main/java/com/p3/persistence/RentalRepository.java
@@ -1,0 +1,9 @@
+package com.p3.persistence;
+
+import com.p3.model.RentalEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RentalRepository extends CrudRepository<RentalEntity, Integer> {
+}
