+++ back/src/main/java/com/p3/persistence/RoleRepository.java
@@ -1,0 +1,12 @@
+package com.p3.persistence;
+
+import com.p3.model.RoleEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface RoleRepository extends CrudRepository<RoleEntity, Integer> {
+    Optional<RoleEntity> findByName(String name);
+}
