+++ back/src/main/java/com/p3/persistence/UserRepository.java
@@ -1,0 +1,13 @@
+package com.p3.persistence;
+
+import com.p3.model.UserEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends CrudRepository<UserEntity, Integer> {
+    Optional<UserEntity> findByEmail(String email);
+    Boolean existsByEmail(String email);
+}
