--- back/src/main/java/controller/P3Application.java
@@ -1,17 +1,0 @@
-package controller;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-@SpringBootApplication(scanBasePackages = {"Api", "Model", "Persistence", "Services", "Security"})
-@EnableJpaRepositories("Persistence")
-@EntityScan("Model")
-public class P3Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(P3Application.class, args);
-	}
-
-}
