+++ back/src/main/resources/application.properties
@@ -1,7 +1,7 @@
-spring.application.name=P3
-spring.datasource.url = jdbc:mariadb://localhost:3307/p3
-spring.datasource.username = root
-spring.datasource.password = root
+spring.application.name=p3
+spring.datasource.url = jdbc:mariadb://localhost:${DB_PORT}/${DB_NAME}
+spring.datasource.username = ${DB_USERNAME}
+spring.datasource.password = ${DB_PASSWORD}
 spring.jpa.hibernate.ddl-auto = update
 server.error.include-message=always
 server.error.include-exception=true
@@ -12,5 +12,5 @@ spring.servlet.multipart.enabled=true
 spring.servlet.multipart.max-file-size=2MB
 spring.servlet.multipart.max-request-size=2MB
 api.jwtExpirationMs=700000000
-assets.savePath=/home/maxdlr/Documents/Github/OC/p3/shared/assets/
-assets.getPath=http://localhost:9090/api/files/file/
+assets.savePath=${ASSETS_PATH}
+assets.getPath=${ASSETS_URL}
