--- back/src/test/java/com/example/back/P3ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.back;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class P3ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
